# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example for response examples value",
        "version": "1.0"
    },
    "paths": {
...
        "/api/cars": {
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "Create new",
                "operationId": "a556a7ef4e7d4fa6f8bac6af3c716ac9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Kia rio 3"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Car"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cars/{carId}": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Get by ID",
                "operationId": "528989c860c5a0e9871808afb2415acc",
                "parameters": [
                    {
                        "name": "carId",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Car"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cars"
                ],
                "summary": "Delete by ID",
                "operationId": "29be6530dbf63e3f4f37688dd8ea7e18",
                "parameters": [
                    {
                        "name": "carId",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully"
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Cars"
                ],
                "summary": "Update by ID",
                "operationId": "f37c0cff7ba37e422d9cb0000228acd7",
                "parameters": [
                    {
                        "name": "carId",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Kia Rio 3"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cars/available": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Get list cars available (actual price)",
                "operationId": "945c3f7a7ae1d0953818bdaad82d55b1",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "allOf": [
                                                    {
                                                        "$ref": "#/components/schemas/Car"
                                                    },
                                                    {
                                                        "properties": {
                                                            "configurations": {
                                                                "type": "array",
                                                                "items": {
                                                                    "allOf": [
                                                                        {
                                                                            "properties": {
                                                                                "id": {
                                                                                    "description": "ID",
                                                                                    "type": "integer",
                                                                                    "example": 10
                                                                                },
                                                                                "name": {
                                                                                    "description": "Name",
                                                                                    "type": "string",
                                                                                    "example": "Lux"
                                                                                },
                                                                                "options": {
                                                                                    "description": "Name",
                                                                                    "type": "array",
                                                                                    "items": {
                                                                                        "type": "string",
                                                                                        "example": "Lanterns"
                                                                                    }
                                                                                },
                                                                                "current_price": {
                                                                                    "description": "Actual price",
                                                                                    "type": "integer",
                                                                                    "example": 10000
                                                                                }
                                                                            },
                                                                            "type": "object"
                                                                        }
                                                                    ]
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Error": {
                "required": [
                    "name"
                ],
                "properties": {
                    "code": {
                        "description": "Code",
                        "type": "integer",
                        "example": 404
                    },
                    "message": {
                        "description": "Message",
                        "type": "string",
                        "example": "Not found"
                    }
                },
                "type": "object"
            },
            "Car": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "ID",
                        "type": "integer",
                        "example": 10
                    },
                    "name": {
                        "description": "Name",
                        "type": "string",
                        "example": "Kia rio 3"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}